Add DataFactoryRegister bean to register KettleDataFactory and SQLStreamingReportDataFactory
